+++ BusinessLogic/DtoModels/RequestDto/RentalReqDto.cs
@@ -9,7 +9,7 @@ namespace BusinessLogic.DtoModels.RequestDto
         public int EmployeeId { get; set; }
         public DateTime DateOfIssue { get; set; }
         public DateTime DateOfRental { get; set; }
-        public double PledgeSum { get; set; }
-        public double? ReturnSum { get; set; }
+        public decimal PledgeSum { get; set; }
+        public decimal? ReturnSum { get; set; }
     }
 }
